Revert "Hands-off tutorial can not be run"
This reverts commit 31a19fe26aac1b27dcc930e384b6c3c92469f9c6.

+++ preseed/start.sh
@@ -123,7 +123,8 @@ then
   rm /var/run/preseed_unspecified_at_boot || true
 fi
 
-if [ -z "$(debconf-get auto-install/classes)" -o \
+if [ -z "$(debconf-get auto-install/classes)" -a \
+     -e /var/run/preseed_unspecified_at_boot  -o \
      -e /var/run/auto-install-had-to-ask-for-preseed ]; then
   db_subst auto-install/classes/title DESC "Which auto-install classes do you want to set?"
   db_settitle auto-install/classes/title
